Rename the server package to reflect what component it has
