+++ kotlin/thrift-example/server/build.gradle.kts
@@ -3,7 +3,7 @@ plugins {
 }
 
 application {
-    mainClassName = "thriftexample.Server"
+    mainClassName = "thriftserver.Server"
 }
 
 dependencies {
